HBASE-30141 Split TestFromClientSide5 (#8193)
(cherry picked from commit f5575e333e03354c6a480797e344226a890c188a)
