+++ hbase-server/src/test/java/org/apache/hadoop/hbase/client/FromClientSideTest5.java
@@ -28,7 +28,6 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
 
 import java.io.IOException;
 import java.time.Duration;
@@ -59,27 +58,14 @@ import org.apache.hadoop.hbase.Waiter;
 import org.apache.hadoop.hbase.client.Scan.ReadType;
 import org.apache.hadoop.hbase.client.metrics.ScanMetrics;
 import org.apache.hadoop.hbase.filter.BinaryComparator;
-import org.apache.hadoop.hbase.filter.Filter;
 import org.apache.hadoop.hbase.filter.FilterList;
 import org.apache.hadoop.hbase.filter.FirstKeyOnlyFilter;
 import org.apache.hadoop.hbase.filter.InclusiveStopFilter;
-import org.apache.hadoop.hbase.filter.KeyOnlyFilter;
-import org.apache.hadoop.hbase.filter.QualifierFilter;
-import org.apache.hadoop.hbase.filter.RegexStringComparator;
 import org.apache.hadoop.hbase.filter.RowFilter;
-import org.apache.hadoop.hbase.filter.SingleColumnValueFilter;
 import org.apache.hadoop.hbase.filter.SubstringComparator;
 import org.apache.hadoop.hbase.filter.ValueFilter;
-import org.apache.hadoop.hbase.io.TimeRange;
 import org.apache.hadoop.hbase.io.hfile.BlockCache;
 import org.apache.hadoop.hbase.io.hfile.CacheConfig;
-import org.apache.hadoop.hbase.ipc.CoprocessorRpcChannel;
-import org.apache.hadoop.hbase.protobuf.ProtobufUtil;
-import org.apache.hadoop.hbase.protobuf.generated.ClientProtos.MutationProto;
-import org.apache.hadoop.hbase.protobuf.generated.ClientProtos.MutationProto.MutationType;
-import org.apache.hadoop.hbase.protobuf.generated.MultiRowMutationProtos.MultiRowMutationService;
-import org.apache.hadoop.hbase.protobuf.generated.MultiRowMutationProtos.MutateRowsRequest;
-import org.apache.hadoop.hbase.protobuf.generated.MultiRowMutationProtos.MutateRowsResponse;
 import org.apache.hadoop.hbase.regionserver.HRegion;
 import org.apache.hadoop.hbase.regionserver.HRegionServer;
 import org.apache.hadoop.hbase.regionserver.HStore;
@@ -211,347 +197,6 @@ public class FromClientSideTest5 extends FromClientSideTestBase {
     scan.addFamily(HConstants.CATALOG_FAMILY);
     try (ResultScanner scanner = table.getScanner(scan)) {
       return scanner.next();
-    }
-  }
-
-  @TestTemplate
-  public void testMultiRowMutation() throws Exception {
-    LOG.info("Starting testMultiRowMutation");
-    final byte[] ROW1 = Bytes.toBytes("testRow1");
-    final byte[] ROW2 = Bytes.toBytes("testRow2");
-    final byte[] ROW3 = Bytes.toBytes("testRow3");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table t = conn.getTable(tableName)) {
-      // Add initial data
-      t.batch(Arrays.asList(new Put(ROW1).addColumn(FAMILY, QUALIFIER, VALUE),
-        new Put(ROW2).addColumn(FAMILY, QUALIFIER, Bytes.toBytes(1L)),
-        new Put(ROW3).addColumn(FAMILY, QUALIFIER, VALUE)), new Object[3]);
-
-      // Execute MultiRowMutation
-      Put put = new Put(ROW).addColumn(FAMILY, QUALIFIER, VALUE);
-      MutationProto m1 = ProtobufUtil.toMutation(MutationType.PUT, put);
-
-      Delete delete = new Delete(ROW1);
-      MutationProto m2 = ProtobufUtil.toMutation(MutationType.DELETE, delete);
-
-      Increment increment = new Increment(ROW2).addColumn(FAMILY, QUALIFIER, 1L);
-      MutationProto m3 = ProtobufUtil.toMutation(MutationType.INCREMENT, increment);
-
-      Append append = new Append(ROW3).addColumn(FAMILY, QUALIFIER, VALUE);
-      MutationProto m4 = ProtobufUtil.toMutation(MutationType.APPEND, append);
-
-      MutateRowsRequest.Builder mrmBuilder = MutateRowsRequest.newBuilder();
-      mrmBuilder.addMutationRequest(m1);
-      mrmBuilder.addMutationRequest(m2);
-      mrmBuilder.addMutationRequest(m3);
-      mrmBuilder.addMutationRequest(m4);
-
-      CoprocessorRpcChannel channel = t.coprocessorService(ROW);
-      MultiRowMutationService.BlockingInterface service =
-        MultiRowMutationService.newBlockingStub(channel);
-      MutateRowsResponse response = service.mutateRows(null, mrmBuilder.build());
-
-      // Assert
-      assertTrue(response.getProcessed());
-
-      Result r = t.get(new Get(ROW));
-      assertEquals(Bytes.toString(VALUE), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-
-      r = t.get(new Get(ROW1));
-      assertTrue(r.isEmpty());
-
-      r = t.get(new Get(ROW2));
-      assertEquals(2L, Bytes.toLong(r.getValue(FAMILY, QUALIFIER)));
-
-      r = t.get(new Get(ROW3));
-      assertEquals(Bytes.toString(VALUE) + Bytes.toString(VALUE),
-        Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-    }
-  }
-
-  @TestTemplate
-  public void testMultiRowMutationWithSingleConditionWhenConditionMatches() throws Exception {
-    final byte[] ROW1 = Bytes.toBytes("testRow1");
-    final byte[] ROW2 = Bytes.toBytes("testRow2");
-    final byte[] VALUE1 = Bytes.toBytes("testValue1");
-    final byte[] VALUE2 = Bytes.toBytes("testValue2");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table t = conn.getTable(tableName)) {
-      // Add initial data
-      t.put(new Put(ROW2).addColumn(FAMILY, QUALIFIER, VALUE2));
-
-      // Execute MultiRowMutation with conditions
-      Put put1 = new Put(ROW).addColumn(FAMILY, QUALIFIER, VALUE);
-      MutationProto m1 = ProtobufUtil.toMutation(MutationType.PUT, put1);
-      Put put2 = new Put(ROW1).addColumn(FAMILY, QUALIFIER, VALUE1);
-      MutationProto m2 = ProtobufUtil.toMutation(MutationType.PUT, put2);
-      Delete delete = new Delete(ROW2);
-      MutationProto m3 = ProtobufUtil.toMutation(MutationType.DELETE, delete);
-
-      MutateRowsRequest.Builder mrmBuilder = MutateRowsRequest.newBuilder();
-      mrmBuilder.addMutationRequest(m1);
-      mrmBuilder.addMutationRequest(m2);
-      mrmBuilder.addMutationRequest(m3);
-      mrmBuilder.addCondition(
-        ProtobufUtil.toCondition(ROW2, FAMILY, QUALIFIER, CompareOperator.EQUAL, VALUE2, null));
-
-      CoprocessorRpcChannel channel = t.coprocessorService(ROW);
-      MultiRowMutationService.BlockingInterface service =
-        MultiRowMutationService.newBlockingStub(channel);
-      MutateRowsResponse response = service.mutateRows(null, mrmBuilder.build());
-
-      // Assert
-      assertTrue(response.getProcessed());
-
-      Result r = t.get(new Get(ROW));
-      assertEquals(Bytes.toString(VALUE), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-
-      r = t.get(new Get(ROW1));
-      assertEquals(Bytes.toString(VALUE1), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-
-      r = t.get(new Get(ROW2));
-      assertTrue(r.isEmpty());
-    }
-  }
-
-  @TestTemplate
-  public void testMultiRowMutationWithSingleConditionWhenConditionNotMatch() throws Exception {
-    final byte[] ROW1 = Bytes.toBytes("testRow1");
-    final byte[] ROW2 = Bytes.toBytes("testRow2");
-    final byte[] VALUE1 = Bytes.toBytes("testValue1");
-    final byte[] VALUE2 = Bytes.toBytes("testValue2");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table t = conn.getTable(tableName)) {
-      // Add initial data
-      t.put(new Put(ROW2).addColumn(FAMILY, QUALIFIER, VALUE2));
-
-      // Execute MultiRowMutation with conditions
-      Put put1 = new Put(ROW).addColumn(FAMILY, QUALIFIER, VALUE);
-      MutationProto m1 = ProtobufUtil.toMutation(MutationType.PUT, put1);
-      Put put2 = new Put(ROW1).addColumn(FAMILY, QUALIFIER, VALUE1);
-      MutationProto m2 = ProtobufUtil.toMutation(MutationType.PUT, put2);
-      Delete delete = new Delete(ROW2);
-      MutationProto m3 = ProtobufUtil.toMutation(MutationType.DELETE, delete);
-
-      MutateRowsRequest.Builder mrmBuilder = MutateRowsRequest.newBuilder();
-      mrmBuilder.addMutationRequest(m1);
-      mrmBuilder.addMutationRequest(m2);
-      mrmBuilder.addMutationRequest(m3);
-      mrmBuilder.addCondition(
-        ProtobufUtil.toCondition(ROW2, FAMILY, QUALIFIER, CompareOperator.EQUAL, VALUE1, null));
-
-      CoprocessorRpcChannel channel = t.coprocessorService(ROW);
-      MultiRowMutationService.BlockingInterface service =
-        MultiRowMutationService.newBlockingStub(channel);
-      MutateRowsResponse response = service.mutateRows(null, mrmBuilder.build());
-
-      // Assert
-      assertFalse(response.getProcessed());
-
-      Result r = t.get(new Get(ROW));
-      assertTrue(r.isEmpty());
-
-      r = t.get(new Get(ROW1));
-      assertTrue(r.isEmpty());
-
-      r = t.get(new Get(ROW2));
-      assertEquals(Bytes.toString(VALUE2), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-    }
-  }
-
-  @TestTemplate
-  public void testMultiRowMutationWithMultipleConditionsWhenConditionsMatch() throws Exception {
-    final byte[] ROW1 = Bytes.toBytes("testRow1");
-    final byte[] ROW2 = Bytes.toBytes("testRow2");
-    final byte[] VALUE1 = Bytes.toBytes("testValue1");
-    final byte[] VALUE2 = Bytes.toBytes("testValue2");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table t = conn.getTable(tableName)) {
-      // Add initial data
-      t.put(new Put(ROW2).addColumn(FAMILY, QUALIFIER, VALUE2));
-
-      // Execute MultiRowMutation with conditions
-      Put put1 = new Put(ROW).addColumn(FAMILY, QUALIFIER, VALUE);
-      MutationProto m1 = ProtobufUtil.toMutation(MutationType.PUT, put1);
-      Put put2 = new Put(ROW1).addColumn(FAMILY, QUALIFIER, VALUE1);
-      MutationProto m2 = ProtobufUtil.toMutation(MutationType.PUT, put2);
-      Delete delete = new Delete(ROW2);
-      MutationProto m3 = ProtobufUtil.toMutation(MutationType.DELETE, delete);
-
-      MutateRowsRequest.Builder mrmBuilder = MutateRowsRequest.newBuilder();
-      mrmBuilder.addMutationRequest(m1);
-      mrmBuilder.addMutationRequest(m2);
-      mrmBuilder.addMutationRequest(m3);
-      mrmBuilder.addCondition(
-        ProtobufUtil.toCondition(ROW, FAMILY, QUALIFIER, CompareOperator.EQUAL, null, null));
-      mrmBuilder.addCondition(
-        ProtobufUtil.toCondition(ROW2, FAMILY, QUALIFIER, CompareOperator.EQUAL, VALUE2, null));
-
-      CoprocessorRpcChannel channel = t.coprocessorService(ROW);
-      MultiRowMutationService.BlockingInterface service =
-        MultiRowMutationService.newBlockingStub(channel);
-      MutateRowsResponse response = service.mutateRows(null, mrmBuilder.build());
-
-      // Assert
-      assertTrue(response.getProcessed());
-
-      Result r = t.get(new Get(ROW));
-      assertEquals(Bytes.toString(VALUE), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-
-      r = t.get(new Get(ROW1));
-      assertEquals(Bytes.toString(VALUE1), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-
-      r = t.get(new Get(ROW2));
-      assertTrue(r.isEmpty());
-    }
-  }
-
-  @TestTemplate
-  public void testMultiRowMutationWithMultipleConditionsWhenConditionsNotMatch() throws Exception {
-    final byte[] ROW1 = Bytes.toBytes("testRow1");
-    final byte[] ROW2 = Bytes.toBytes("testRow2");
-    final byte[] VALUE1 = Bytes.toBytes("testValue1");
-    final byte[] VALUE2 = Bytes.toBytes("testValue2");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table t = conn.getTable(tableName)) {
-      // Add initial data
-      t.put(new Put(ROW2).addColumn(FAMILY, QUALIFIER, VALUE2));
-
-      // Execute MultiRowMutation with conditions
-      Put put1 = new Put(ROW).addColumn(FAMILY, QUALIFIER, VALUE);
-      MutationProto m1 = ProtobufUtil.toMutation(MutationType.PUT, put1);
-      Put put2 = new Put(ROW1).addColumn(FAMILY, QUALIFIER, VALUE1);
-      MutationProto m2 = ProtobufUtil.toMutation(MutationType.PUT, put2);
-      Delete delete = new Delete(ROW2);
-      MutationProto m3 = ProtobufUtil.toMutation(MutationType.DELETE, delete);
-
-      MutateRowsRequest.Builder mrmBuilder = MutateRowsRequest.newBuilder();
-      mrmBuilder.addMutationRequest(m1);
-      mrmBuilder.addMutationRequest(m2);
-      mrmBuilder.addMutationRequest(m3);
-      mrmBuilder.addCondition(
-        ProtobufUtil.toCondition(ROW1, FAMILY, QUALIFIER, CompareOperator.EQUAL, null, null));
-      mrmBuilder.addCondition(
-        ProtobufUtil.toCondition(ROW2, FAMILY, QUALIFIER, CompareOperator.EQUAL, VALUE1, null));
-
-      CoprocessorRpcChannel channel = t.coprocessorService(ROW);
-      MultiRowMutationService.BlockingInterface service =
-        MultiRowMutationService.newBlockingStub(channel);
-      MutateRowsResponse response = service.mutateRows(null, mrmBuilder.build());
-
-      // Assert
-      assertFalse(response.getProcessed());
-
-      Result r = t.get(new Get(ROW));
-      assertTrue(r.isEmpty());
-
-      r = t.get(new Get(ROW1));
-      assertTrue(r.isEmpty());
-
-      r = t.get(new Get(ROW2));
-      assertEquals(Bytes.toString(VALUE2), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-    }
-  }
-
-  @TestTemplate
-  public void testMultiRowMutationWithFilterConditionWhenConditionMatches() throws Exception {
-    final byte[] ROW1 = Bytes.toBytes("testRow1");
-    final byte[] ROW2 = Bytes.toBytes("testRow2");
-    final byte[] QUALIFIER2 = Bytes.toBytes("testQualifier2");
-    final byte[] VALUE1 = Bytes.toBytes("testValue1");
-    final byte[] VALUE2 = Bytes.toBytes("testValue2");
-    final byte[] VALUE3 = Bytes.toBytes("testValue3");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table t = conn.getTable(tableName)) {
-      // Add initial data
-      t.put(
-        new Put(ROW2).addColumn(FAMILY, QUALIFIER, VALUE2).addColumn(FAMILY, QUALIFIER2, VALUE3));
-
-      // Execute MultiRowMutation with conditions
-      Put put1 = new Put(ROW).addColumn(FAMILY, QUALIFIER, VALUE);
-      MutationProto m1 = ProtobufUtil.toMutation(MutationType.PUT, put1);
-      Put put2 = new Put(ROW1).addColumn(FAMILY, QUALIFIER, VALUE1);
-      MutationProto m2 = ProtobufUtil.toMutation(MutationType.PUT, put2);
-      Delete delete = new Delete(ROW2);
-      MutationProto m3 = ProtobufUtil.toMutation(MutationType.DELETE, delete);
-
-      MutateRowsRequest.Builder mrmBuilder = MutateRowsRequest.newBuilder();
-      mrmBuilder.addMutationRequest(m1);
-      mrmBuilder.addMutationRequest(m2);
-      mrmBuilder.addMutationRequest(m3);
-      mrmBuilder.addCondition(ProtobufUtil.toCondition(ROW2,
-        new FilterList(
-          new SingleColumnValueFilter(FAMILY, QUALIFIER, CompareOperator.EQUAL, VALUE2),
-          new SingleColumnValueFilter(FAMILY, QUALIFIER2, CompareOperator.EQUAL, VALUE3)),
-        null));
-
-      CoprocessorRpcChannel channel = t.coprocessorService(ROW);
-      MultiRowMutationService.BlockingInterface service =
-        MultiRowMutationService.newBlockingStub(channel);
-      MutateRowsResponse response = service.mutateRows(null, mrmBuilder.build());
-
-      // Assert
-      assertTrue(response.getProcessed());
-
-      Result r = t.get(new Get(ROW));
-      assertEquals(Bytes.toString(VALUE), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-
-      r = t.get(new Get(ROW1));
-      assertEquals(Bytes.toString(VALUE1), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
-
-      r = t.get(new Get(ROW2));
-      assertTrue(r.isEmpty());
-    }
-  }
-
-  @TestTemplate
-  public void testMultiRowMutationWithFilterConditionWhenConditionNotMatch() throws Exception {
-    final byte[] ROW1 = Bytes.toBytes("testRow1");
-    final byte[] ROW2 = Bytes.toBytes("testRow2");
-    final byte[] QUALIFIER2 = Bytes.toBytes("testQualifier2");
-    final byte[] VALUE1 = Bytes.toBytes("testValue1");
-    final byte[] VALUE2 = Bytes.toBytes("testValue2");
-    final byte[] VALUE3 = Bytes.toBytes("testValue3");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table t = conn.getTable(tableName)) {
-      // Add initial data
-      t.put(
-        new Put(ROW2).addColumn(FAMILY, QUALIFIER, VALUE2).addColumn(FAMILY, QUALIFIER2, VALUE3));
-
-      // Execute MultiRowMutation with conditions
-      Put put1 = new Put(ROW).addColumn(FAMILY, QUALIFIER, VALUE);
-      MutationProto m1 = ProtobufUtil.toMutation(MutationType.PUT, put1);
-      Put put2 = new Put(ROW1).addColumn(FAMILY, QUALIFIER, VALUE1);
-      MutationProto m2 = ProtobufUtil.toMutation(MutationType.PUT, put2);
-      Delete delete = new Delete(ROW2);
-      MutationProto m3 = ProtobufUtil.toMutation(MutationType.DELETE, delete);
-
-      MutateRowsRequest.Builder mrmBuilder = MutateRowsRequest.newBuilder();
-      mrmBuilder.addMutationRequest(m1);
-      mrmBuilder.addMutationRequest(m2);
-      mrmBuilder.addMutationRequest(m3);
-      mrmBuilder.addCondition(ProtobufUtil.toCondition(ROW2,
-        new FilterList(
-          new SingleColumnValueFilter(FAMILY, QUALIFIER, CompareOperator.EQUAL, VALUE2),
-          new SingleColumnValueFilter(FAMILY, QUALIFIER2, CompareOperator.EQUAL, VALUE2)),
-        null));
-
-      CoprocessorRpcChannel channel = t.coprocessorService(ROW);
-      MultiRowMutationService.BlockingInterface service =
-        MultiRowMutationService.newBlockingStub(channel);
-      MutateRowsResponse response = service.mutateRows(null, mrmBuilder.build());
-
-      // Assert
-      assertFalse(response.getProcessed());
-
-      Result r = t.get(new Get(ROW));
-      assertTrue(r.isEmpty());
-
-      r = t.get(new Get(ROW1));
-      assertTrue(r.isEmpty());
-
-      r = t.get(new Get(ROW2));
-      assertEquals(Bytes.toString(VALUE2), Bytes.toString(r.getValue(FAMILY, QUALIFIER)));
     }
   }
 
@@ -858,305 +503,6 @@ public class FromClientSideTest5 extends FromClientSideTestBase {
       }
       executorService.shutdownNow();
       assertNull(error.get());
-    }
-  }
-
-  @TestTemplate
-  public void testCheckAndPut() throws IOException {
-    final byte[] anotherrow = Bytes.toBytes("anotherrow");
-    final byte[] value2 = Bytes.toBytes("abcd");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table table = conn.getTable(tableName)) {
-      Put put1 = new Put(ROW);
-      put1.addColumn(FAMILY, QUALIFIER, VALUE);
-
-      // row doesn't exist, so using non-null value should be considered "not match".
-      boolean ok =
-        table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).ifEquals(VALUE).thenPut(put1);
-      assertFalse(ok);
-
-      // row doesn't exist, so using "ifNotExists" should be considered "match".
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).ifNotExists().thenPut(put1);
-      assertTrue(ok);
-
-      // row now exists, so using "ifNotExists" should be considered "not match".
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).ifNotExists().thenPut(put1);
-      assertFalse(ok);
-
-      Put put2 = new Put(ROW);
-      put2.addColumn(FAMILY, QUALIFIER, value2);
-
-      // row now exists, use the matching value to check
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).ifEquals(VALUE).thenPut(put2);
-      assertTrue(ok);
-
-      Put put3 = new Put(anotherrow);
-      put3.addColumn(FAMILY, QUALIFIER, VALUE);
-
-      // try to do CheckAndPut on different rows
-      try {
-        table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).ifEquals(value2).thenPut(put3);
-        fail("trying to check and modify different rows should have failed.");
-      } catch (Exception ignored) {
-      }
-    }
-  }
-
-  @TestTemplate
-  public void testCheckAndMutateWithTimeRange() throws IOException {
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table table = conn.getTable(tableName)) {
-      final long ts = EnvironmentEdgeManager.currentTime() / 2;
-      Put put = new Put(ROW);
-      put.addColumn(FAMILY, QUALIFIER, ts, VALUE);
-
-      boolean ok =
-        table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).ifNotExists().thenPut(put);
-      assertTrue(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .timeRange(TimeRange.at(ts + 10000)).ifEquals(VALUE).thenPut(put);
-      assertFalse(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .timeRange(TimeRange.from(ts + 10000)).ifEquals(VALUE).thenPut(put);
-      assertFalse(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .timeRange(TimeRange.between(ts + 10000, ts + 20000)).ifEquals(VALUE).thenPut(put);
-      assertFalse(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).timeRange(TimeRange.until(ts))
-        .ifEquals(VALUE).thenPut(put);
-      assertFalse(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).timeRange(TimeRange.at(ts))
-        .ifEquals(VALUE).thenPut(put);
-      assertTrue(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).timeRange(TimeRange.from(ts))
-        .ifEquals(VALUE).thenPut(put);
-      assertTrue(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .timeRange(TimeRange.between(ts, ts + 20000)).ifEquals(VALUE).thenPut(put);
-      assertTrue(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .timeRange(TimeRange.until(ts + 10000)).ifEquals(VALUE).thenPut(put);
-      assertTrue(ok);
-
-      RowMutations rm = new RowMutations(ROW).add((Mutation) put);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .timeRange(TimeRange.at(ts + 10000)).ifEquals(VALUE).thenMutate(rm);
-      assertFalse(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).timeRange(TimeRange.at(ts))
-        .ifEquals(VALUE).thenMutate(rm);
-      assertTrue(ok);
-
-      Delete delete = new Delete(ROW).addColumn(FAMILY, QUALIFIER);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .timeRange(TimeRange.at(ts + 10000)).ifEquals(VALUE).thenDelete(delete);
-      assertFalse(ok);
-
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).timeRange(TimeRange.at(ts))
-        .ifEquals(VALUE).thenDelete(delete);
-      assertTrue(ok);
-    }
-  }
-
-  @TestTemplate
-  public void testCheckAndPutWithCompareOp() throws IOException {
-    final byte[] value1 = Bytes.toBytes("aaaa");
-    final byte[] value2 = Bytes.toBytes("bbbb");
-    final byte[] value3 = Bytes.toBytes("cccc");
-    final byte[] value4 = Bytes.toBytes("dddd");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table table = conn.getTable(tableName)) {
-      Put put2 = new Put(ROW);
-      put2.addColumn(FAMILY, QUALIFIER, value2);
-
-      Put put3 = new Put(ROW);
-      put3.addColumn(FAMILY, QUALIFIER, value3);
-
-      // row doesn't exist, so using "ifNotExists" should be considered "match".
-      boolean ok =
-        table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).ifNotExists().thenPut(put2);
-      assertTrue(ok);
-
-      // cell = "bbbb", using "aaaa" to compare only LESS/LESS_OR_EQUAL/NOT_EQUAL
-      // turns out "match"
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER, value1).thenPut(put2);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.EQUAL, value1).thenPut(put2);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER_OR_EQUAL, value1).thenPut(put2);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS, value1).thenPut(put2);
-      assertTrue(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS_OR_EQUAL, value1).thenPut(put2);
-      assertTrue(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.NOT_EQUAL, value1).thenPut(put3);
-      assertTrue(ok);
-
-      // cell = "cccc", using "dddd" to compare only LARGER/LARGER_OR_EQUAL/NOT_EQUAL
-      // turns out "match"
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS, value4).thenPut(put3);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS_OR_EQUAL, value4).thenPut(put3);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.EQUAL, value4).thenPut(put3);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER, value4).thenPut(put3);
-      assertTrue(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER_OR_EQUAL, value4).thenPut(put3);
-      assertTrue(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.NOT_EQUAL, value4).thenPut(put2);
-      assertTrue(ok);
-
-      // cell = "bbbb", using "bbbb" to compare only GREATER_OR_EQUAL/LESS_OR_EQUAL/EQUAL
-      // turns out "match"
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER, value2).thenPut(put2);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.NOT_EQUAL, value2).thenPut(put2);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS, value2).thenPut(put2);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER_OR_EQUAL, value2).thenPut(put2);
-      assertTrue(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS_OR_EQUAL, value2).thenPut(put2);
-      assertTrue(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.EQUAL, value2).thenPut(put3);
-      assertTrue(ok);
-    }
-  }
-
-  @TestTemplate
-  public void testCheckAndDelete() throws IOException {
-    final byte[] value1 = Bytes.toBytes("aaaa");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table table = conn.getTable(tableName)) {
-      Put put = new Put(ROW);
-      put.addColumn(FAMILY, QUALIFIER, value1);
-      table.put(put);
-
-      Delete delete = new Delete(ROW);
-      delete.addColumns(FAMILY, QUALIFIER);
-
-      boolean ok =
-        table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER).ifEquals(value1).thenDelete(delete);
-      assertTrue(ok);
-    }
-  }
-
-  @TestTemplate
-  public void testCheckAndDeleteWithCompareOp() throws IOException {
-    final byte[] value1 = Bytes.toBytes("aaaa");
-    final byte[] value2 = Bytes.toBytes("bbbb");
-    final byte[] value3 = Bytes.toBytes("cccc");
-    final byte[] value4 = Bytes.toBytes("dddd");
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table table = conn.getTable(tableName)) {
-      Put put2 = new Put(ROW);
-      put2.addColumn(FAMILY, QUALIFIER, value2);
-      table.put(put2);
-
-      Put put3 = new Put(ROW);
-      put3.addColumn(FAMILY, QUALIFIER, value3);
-
-      Delete delete = new Delete(ROW);
-      delete.addColumns(FAMILY, QUALIFIER);
-
-      // cell = "bbbb", using "aaaa" to compare only LESS/LESS_OR_EQUAL/NOT_EQUAL
-      // turns out "match"
-      boolean ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER, value1).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.EQUAL, value1).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER_OR_EQUAL, value1).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS, value1).thenDelete(delete);
-      assertTrue(ok);
-      table.put(put2);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS_OR_EQUAL, value1).thenDelete(delete);
-      assertTrue(ok);
-      table.put(put2);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.NOT_EQUAL, value1).thenDelete(delete);
-      assertTrue(ok);
-
-      // cell = "cccc", using "dddd" to compare only LARGER/LARGER_OR_EQUAL/NOT_EQUAL
-      // turns out "match"
-      table.put(put3);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS, value4).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS_OR_EQUAL, value4).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.EQUAL, value4).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER, value4).thenDelete(delete);
-      assertTrue(ok);
-      table.put(put3);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER_OR_EQUAL, value4).thenDelete(delete);
-      assertTrue(ok);
-      table.put(put3);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.NOT_EQUAL, value4).thenDelete(delete);
-      assertTrue(ok);
-
-      // cell = "bbbb", using "bbbb" to compare only GREATER_OR_EQUAL/LESS_OR_EQUAL/EQUAL
-      // turns out "match"
-      table.put(put2);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER, value2).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.NOT_EQUAL, value2).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS, value2).thenDelete(delete);
-      assertFalse(ok);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.GREATER_OR_EQUAL, value2).thenDelete(delete);
-      assertTrue(ok);
-      table.put(put2);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.LESS_OR_EQUAL, value2).thenDelete(delete);
-      assertTrue(ok);
-      table.put(put2);
-      ok = table.checkAndMutate(ROW, FAMILY).qualifier(QUALIFIER)
-        .ifMatches(CompareOperator.EQUAL, value2).thenDelete(delete);
-      assertTrue(ok);
     }
   }
 
@@ -1749,631 +1095,6 @@ public class FromClientSideTest5 extends FromClientSideTestBase {
         }
         assertEquals(insertNum, count);
       }
-    }
-  }
-
-  @TestTemplate
-  public void testSuperSimpleWithReverseScan() throws Exception {
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table ht = conn.getTable(tableName)) {
-      Put put = new Put(Bytes.toBytes("0-b11111-0000000000000000000"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b11111-0000000000000000002"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b11111-0000000000000000004"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b11111-0000000000000000006"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b11111-0000000000000000008"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b22222-0000000000000000001"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b22222-0000000000000000003"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b22222-0000000000000000005"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b22222-0000000000000000007"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      put = new Put(Bytes.toBytes("0-b22222-0000000000000000009"));
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-      Scan scan = new Scan().withStartRow(Bytes.toBytes("0-b11111-9223372036854775807"))
-        .withStopRow(Bytes.toBytes("0-b11111-0000000000000000000"), true);
-      scan.setReversed(true);
-      try (ResultScanner scanner = ht.getScanner(scan)) {
-        Result result = scanner.next();
-        assertTrue(Bytes.equals(result.getRow(), Bytes.toBytes("0-b11111-0000000000000000008")));
-      }
-    }
-  }
-
-  @TestTemplate
-  public void testFiltersWithReverseScan() throws Exception {
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table ht = conn.getTable(tableName)) {
-      byte[][] ROWS = makeN(ROW, 10);
-      byte[][] QUALIFIERS =
-        { Bytes.toBytes("col0-<d2v1>-<d3v2>"), Bytes.toBytes("col1-<d2v1>-<d3v2>"),
-          Bytes.toBytes("col2-<d2v1>-<d3v2>"), Bytes.toBytes("col3-<d2v1>-<d3v2>"),
-          Bytes.toBytes("col4-<d2v1>-<d3v2>"), Bytes.toBytes("col5-<d2v1>-<d3v2>"),
-          Bytes.toBytes("col6-<d2v1>-<d3v2>"), Bytes.toBytes("col7-<d2v1>-<d3v2>"),
-          Bytes.toBytes("col8-<d2v1>-<d3v2>"), Bytes.toBytes("col9-<d2v1>-<d3v2>") };
-      for (int i = 0; i < 10; i++) {
-        Put put = new Put(ROWS[i]);
-        put.addColumn(FAMILY, QUALIFIERS[i], VALUE);
-        ht.put(put);
-      }
-      Scan scan = new Scan();
-      scan.setReversed(true);
-      scan.addFamily(FAMILY);
-      Filter filter =
-        new QualifierFilter(CompareOperator.EQUAL, new RegexStringComparator("col[1-5]"));
-      scan.setFilter(filter);
-      try (ResultScanner scanner = ht.getScanner(scan)) {
-        int expectedIndex = 5;
-        for (Result result : scanner) {
-          assertEquals(1, result.size());
-          Cell c = result.rawCells()[0];
-          assertTrue(Bytes.equals(c.getRowArray(), c.getRowOffset(), c.getRowLength(),
-            ROWS[expectedIndex], 0, ROWS[expectedIndex].length));
-          assertTrue(
-            Bytes.equals(c.getQualifierArray(), c.getQualifierOffset(), c.getQualifierLength(),
-              QUALIFIERS[expectedIndex], 0, QUALIFIERS[expectedIndex].length));
-          expectedIndex--;
-        }
-        assertEquals(0, expectedIndex);
-      }
-    }
-  }
-
-  @TestTemplate
-  public void testKeyOnlyFilterWithReverseScan() throws Exception {
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table ht = conn.getTable(tableName)) {
-      byte[][] ROWS = makeN(ROW, 10);
-      byte[][] QUALIFIERS =
-        { Bytes.toBytes("col0-<d2v1>-<d3v2>"), Bytes.toBytes("col1-<d2v1>-<d3v2>"),
-          Bytes.toBytes("col2-<d2v1>-<d3v2>"), Bytes.toBytes("col3-<d2v1>-<d3v2>"),
-          Bytes.toBytes("col4-<d2v1>-<d3v2>"), Bytes.toBytes("col5-<d2v1>-<d3v2>"),
-          Bytes.toBytes("col6-<d2v1>-<d3v2>"), Bytes.toBytes("col7-<d2v1>-<d3v2>"),
-          Bytes.toBytes("col8-<d2v1>-<d3v2>"), Bytes.toBytes("col9-<d2v1>-<d3v2>") };
-      for (int i = 0; i < 10; i++) {
-        Put put = new Put(ROWS[i]);
-        put.addColumn(FAMILY, QUALIFIERS[i], VALUE);
-        ht.put(put);
-      }
-      Scan scan = new Scan();
-      scan.setReversed(true);
-      scan.addFamily(FAMILY);
-      Filter filter = new KeyOnlyFilter(true);
-      scan.setFilter(filter);
-      try (ResultScanner ignored = ht.getScanner(scan)) {
-        int count = 0;
-        for (Result result : ht.getScanner(scan)) {
-          assertEquals(1, result.size());
-          assertEquals(Bytes.SIZEOF_INT, result.rawCells()[0].getValueLength());
-          assertEquals(VALUE.length, Bytes.toInt(CellUtil.cloneValue(result.rawCells()[0])));
-          count++;
-        }
-        assertEquals(10, count);
-      }
-    }
-  }
-
-  /**
-   * Test simple table and non-existent row cases.
-   */
-  @TestTemplate
-  public void testSimpleMissingWithReverseScan() throws Exception {
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table ht = conn.getTable(tableName)) {
-      byte[][] ROWS = makeN(ROW, 4);
-
-      // Try to get a row on an empty table
-      Scan scan = new Scan();
-      scan.setReversed(true);
-      Result result = getSingleScanResult(ht, scan);
-      assertNullResult(result);
-
-      scan = new Scan().withStartRow(ROWS[0]);
-      scan.setReversed(true);
-      result = getSingleScanResult(ht, scan);
-      assertNullResult(result);
-
-      scan = new Scan().withStartRow(ROWS[0]).withStopRow(ROWS[1], true);
-      scan.setReversed(true);
-      result = getSingleScanResult(ht, scan);
-      assertNullResult(result);
-
-      scan = new Scan();
-      scan.setReversed(true);
-      scan.addFamily(FAMILY);
-      result = getSingleScanResult(ht, scan);
-      assertNullResult(result);
-
-      scan = new Scan();
-      scan.setReversed(true);
-      scan.addColumn(FAMILY, QUALIFIER);
-      result = getSingleScanResult(ht, scan);
-      assertNullResult(result);
-
-      // Insert a row
-      Put put = new Put(ROWS[2]);
-      put.addColumn(FAMILY, QUALIFIER, VALUE);
-      ht.put(put);
-
-      // Make sure we can scan the row
-      scan = new Scan();
-      scan.setReversed(true);
-      result = getSingleScanResult(ht, scan);
-      assertSingleResult(result, ROWS[2], FAMILY, QUALIFIER, VALUE);
-
-      scan = new Scan().withStartRow(ROWS[3]).withStopRow(ROWS[0], true);
-      scan.setReversed(true);
-      result = getSingleScanResult(ht, scan);
-      assertSingleResult(result, ROWS[2], FAMILY, QUALIFIER, VALUE);
-
-      scan = new Scan().withStartRow(ROWS[2]).withStopRow(ROWS[1], true);
-      scan.setReversed(true);
-      result = getSingleScanResult(ht, scan);
-      assertSingleResult(result, ROWS[2], FAMILY, QUALIFIER, VALUE);
-
-      // Try to scan empty rows around it
-      // Introduced MemStore#shouldSeekForReverseScan to fix the following
-      scan = new Scan().withStartRow(ROWS[1]);
-      scan.setReversed(true);
-      result = getSingleScanResult(ht, scan);
-      assertNullResult(result);
-    }
-  }
-
-  @TestTemplate
-  public void testNullWithReverseScan() throws Exception {
-    TEST_UTIL.createTable(tableName, FAMILY);
-    try (Connection conn = getConnection(); Table ht = conn.getTable(tableName)) {
-      // Null qualifier (should work)
-      Put put = new Put(ROW);
-      put.addColumn(FAMILY, null, VALUE);
-      ht.put(put);
-      scanTestNull(ht, ROW, FAMILY, VALUE, true);
-      Delete delete = new Delete(ROW);
-      delete.addColumns(FAMILY, null);
-      ht.delete(delete);
-    }
-
-    // Use a new table
-    TableName newTableName = TableName.valueOf(tableName.toString() + "2");
-    TEST_UTIL.createTable(newTableName, FAMILY);
-    try (Connection conn = getConnection(); Table ht = conn.getTable(newTableName)) {
-      // Empty qualifier, byte[0] instead of null (should work)
-      Put put = new Put(ROW);
-      put.addColumn(FAMILY, HConstants.EMPTY_BYTE_ARRAY, VALUE);
-      ht.put(put);
-      scanTestNull(ht, ROW, FAMILY, VALUE, true);
-      TEST_UTIL.flush();
-      scanTestNull(ht, ROW, FAMILY, VALUE, true);
-      Delete delete = new Delete(ROW);
-      delete.addColumns(FAMILY, HConstants.EMPTY_BYTE_ARRAY);
-      ht.delete(delete);
-      // Null value
-      put = new Put(ROW);
-      put.addColumn(FAMILY, QUALIFIER, null);
-      ht.put(put);
-      Scan scan = new Scan();
-      scan.setReversed(true);
-      scan.addColumn(FAMILY, QUALIFIER);
-      Result result = getSingleScanResult(ht, scan);
-      assertSingleResult(result, ROW, FAMILY, QUALIFIER, null);
-    }
-  }
-
-  @TestTemplate
-  @SuppressWarnings("checkstyle:MethodLength")
-  public void testDeletesWithReverseScan() throws Exception {
-    byte[][] ROWS = makeNAscii(ROW, 6);
-    byte[][] FAMILIES = makeNAscii(FAMILY, 3);
-    byte[][] VALUES = makeN(VALUE, 5);
-    long[] ts = { 1000, 2000, 3000, 4000, 5000 };
-    TEST_UTIL.createTable(tableName, FAMILIES, 3);
-    try (Connection conn = getConnection(); Table ht = conn.getTable(tableName)) {
-      Put put = new Put(ROW);
-      put.addColumn(FAMILIES[0], QUALIFIER, ts[0], VALUES[0]);
-      put.addColumn(FAMILIES[0], QUALIFIER, ts[1], VALUES[1]);
-      ht.put(put);
-
-      Delete delete = new Delete(ROW);
-      delete.addFamily(FAMILIES[0], ts[0]);
-      ht.delete(delete);
-
-      Scan scan = new Scan().withStartRow(ROW);
-      scan.setReversed(true);
-      scan.addFamily(FAMILIES[0]);
-      scan.readVersions(Integer.MAX_VALUE);
-      Result result = getSingleScanResult(ht, scan);
-      assertNResult(result, ROW, FAMILIES[0], QUALIFIER, new long[] { ts[1] },
-        new byte[][] { VALUES[1] }, 0, 0);
-
-      // Test delete latest version
-      put = new Put(ROW);
-      put.addColumn(FAMILIES[0], QUALIFIER, ts[4], VALUES[4]);
-      put.addColumn(FAMILIES[0], QUALIFIER, ts[2], VALUES[2]);
-      put.addColumn(FAMILIES[0], QUALIFIER, ts[3], VALUES[3]);
-      put.addColumn(FAMILIES[0], null, ts[4], VALUES[4]);
-      put.addColumn(FAMILIES[0], null, ts[2], VALUES[2]);
-      put.addColumn(FAMILIES[0], null, ts[3], VALUES[3]);
-      ht.put(put);
-
-      delete = new Delete(ROW);
-      delete.addColumn(FAMILIES[0], QUALIFIER); // ts[4]
-      ht.delete(delete);
-
-      scan = new Scan().withStartRow(ROW);
-      scan.setReversed(true);
-      scan.addColumn(FAMILIES[0], QUALIFIER);
-      scan.readVersions(Integer.MAX_VALUE);
-      result = getSingleScanResult(ht, scan);
-      assertNResult(result, ROW, FAMILIES[0], QUALIFIER, new long[] { ts[1], ts[2], ts[3] },
-        new byte[][] { VALUES[1], VALUES[2], VALUES[3] }, 0, 2);
-
-      // Test for HBASE-1847
-      delete = new Delete(ROW);
-      delete.addColumn(FAMILIES[0], null);
-      ht.delete(delete);
-
-      // Cleanup null qualifier
-      delete = new Delete(ROW);
-      delete.addColumns(FAMILIES[0], null);
-      ht.delete(delete);
-
-      // Expected client behavior might be that you can re-put deleted values
-      // But alas, this is not to be. We can't put them back in either case.
-
-      put = new Put(ROW);
-      put.addColumn(FAMILIES[0], QUALIFIER, ts[0], VALUES[0]);
-      put.addColumn(FAMILIES[0], QUALIFIER, ts[4], VALUES[4]);
-      ht.put(put);
-
-      // The Scanner returns the previous values, the expected-naive-unexpected
-      // behavior
-
-      scan = new Scan().withStartRow(ROW);
-      scan.setReversed(true);
-      scan.addFamily(FAMILIES[0]);
-      scan.readVersions(Integer.MAX_VALUE);
-      result = getSingleScanResult(ht, scan);
-      assertNResult(result, ROW, FAMILIES[0], QUALIFIER, new long[] { ts[1], ts[2], ts[3] },
-        new byte[][] { VALUES[1], VALUES[2], VALUES[3] }, 0, 2);
-
-      // Test deleting an entire family from one row but not the other various
-      // ways
-
-      put = new Put(ROWS[0]);
-      put.addColumn(FAMILIES[1], QUALIFIER, ts[0], VALUES[0]);
-      put.addColumn(FAMILIES[1], QUALIFIER, ts[1], VALUES[1]);
-      put.addColumn(FAMILIES[2], QUALIFIER, ts[2], VALUES[2]);
-      put.addColumn(FAMILIES[2], QUALIFIER, ts[3], VALUES[3]);
-      ht.put(put);
-
-      put = new Put(ROWS[1]);
-      put.addColumn(FAMILIES[1], QUALIFIER, ts[0], VALUES[0]);
-      put.addColumn(FAMILIES[1], QUALIFIER, ts[1], VALUES[1]);
-      put.addColumn(FAMILIES[2], QUALIFIER, ts[2], VALUES[2]);
-      put.addColumn(FAMILIES[2], QUALIFIER, ts[3], VALUES[3]);
-      ht.put(put);
-
-      put = new Put(ROWS[2]);
-      put.addColumn(FAMILIES[1], QUALIFIER, ts[0], VALUES[0]);
-      put.addColumn(FAMILIES[1], QUALIFIER, ts[1], VALUES[1]);
-      put.addColumn(FAMILIES[2], QUALIFIER, ts[2], VALUES[2]);
-      put.addColumn(FAMILIES[2], QUALIFIER, ts[3], VALUES[3]);
-      ht.put(put);
-
-      delete = new Delete(ROWS[0]);
-      delete.addFamily(FAMILIES[2]);
-      ht.delete(delete);
-
-      delete = new Delete(ROWS[1]);
-      delete.addColumns(FAMILIES[1], QUALIFIER);
-      ht.delete(delete);
-
-      delete = new Delete(ROWS[2]);
-      delete.addColumn(FAMILIES[1], QUALIFIER);
-      delete.addColumn(FAMILIES[1], QUALIFIER);
-      delete.addColumn(FAMILIES[2], QUALIFIER);
-      ht.delete(delete);
-
-      scan = new Scan().withStartRow(ROWS[0]);
-      scan.setReversed(true);
-      scan.addFamily(FAMILIES[1]);
-      scan.addFamily(FAMILIES[2]);
-      scan.readVersions(Integer.MAX_VALUE);
-      result = getSingleScanResult(ht, scan);
-      assertEquals(2, result.size(), "Expected 2 keys but received " + result.size());
-      assertNResult(result, ROWS[0], FAMILIES[1], QUALIFIER, new long[] { ts[0], ts[1] },
-        new byte[][] { VALUES[0], VALUES[1] }, 0, 1);
-
-      scan = new Scan().withStartRow(ROWS[1]);
-      scan.setReversed(true);
-      scan.addFamily(FAMILIES[1]);
-      scan.addFamily(FAMILIES[2]);
-      scan.readVersions(Integer.MAX_VALUE);
-      result = getSingleScanResult(ht, scan);
-      assertEquals(2, result.size(), "Expected 2 keys but received " + result.size());
-
-      scan = new Scan().withStartRow(ROWS[2]);
-      scan.setReversed(true);
-      scan.addFamily(FAMILIES[1]);
-      scan.addFamily(FAMILIES[2]);
-      scan.readVersions(Integer.MAX_VALUE);
-      result = getSingleScanResult(ht, scan);
-      assertEquals(1, result.size());
-      assertNResult(result, ROWS[2], FAMILIES[2], QUALIFIER, new long[] { ts[2] },
-        new byte[][] { VALUES[2] }, 0, 0);
-
-      // Test if we delete the family first in one row (HBASE-1541)
-
-      delete = new Delete(ROWS[3]);
-      delete.addFamily(FAMILIES[1]);
-      ht.delete(delete);
-
-      put = new Put(ROWS[3]);
-      put.addColumn(FAMILIES[2], QUALIFIER, VALUES[0]);
-      ht.put(put);
-
-      put = new Put(ROWS[4]);
-      put.addColumn(FAMILIES[1], QUALIFIER, VALUES[1]);
-      put.addColumn(FAMILIES[2], QUALIFIER, VALUES[2]);
-      ht.put(put);
-
-      scan = new Scan().withStartRow(ROWS[4]);
-      scan.setReversed(true);
-      scan.addFamily(FAMILIES[1]);
-      scan.addFamily(FAMILIES[2]);
-      scan.readVersions(Integer.MAX_VALUE);
-      try (ResultScanner scanner = ht.getScanner(scan)) {
-        result = scanner.next();
-        assertEquals(2, result.size(), "Expected 2 keys but received " + result.size());
-        assertTrue(Bytes.equals(CellUtil.cloneRow(result.rawCells()[0]), ROWS[4]));
-        assertTrue(Bytes.equals(CellUtil.cloneRow(result.rawCells()[1]), ROWS[4]));
-        assertTrue(Bytes.equals(CellUtil.cloneValue(result.rawCells()[0]), VALUES[1]));
-        assertTrue(Bytes.equals(CellUtil.cloneValue(result.rawCells()[1]), VALUES[2]));
-        result = scanner.next();
-        assertEquals(1, result.size(), "Expected 1 key but received " + result.size());
-        assertTrue(Bytes.equals(CellUtil.cloneRow(result.rawCells()[0]), ROWS[3]));
-        assertTrue(Bytes.equals(CellUtil.cloneValue(result.rawCells()[0]), VALUES[0]));
-      }
-    }
-  }
-
-  /**
-   * Tests reversed scan under multi regions
-   */
-  @TestTemplate
-  public void testReversedScanUnderMultiRegions() throws Exception {
-    // Test Initialization.
-    byte[] maxByteArray = ConnectionUtils.MAX_BYTE_ARRAY;
-    byte[][] splitRows = new byte[][] { Bytes.toBytes("005"),
-      Bytes.add(Bytes.toBytes("005"), Bytes.multiple(maxByteArray, 16)), Bytes.toBytes("006"),
-      Bytes.add(Bytes.toBytes("006"), Bytes.multiple(maxByteArray, 8)), Bytes.toBytes("007"),
-      Bytes.add(Bytes.toBytes("007"), Bytes.multiple(maxByteArray, 4)), Bytes.toBytes("008"),
-      Bytes.multiple(maxByteArray, 2) };
-    TEST_UTIL.createTable(tableName, FAMILY, splitRows);
-    TEST_UTIL.waitUntilAllRegionsAssigned(tableName);
-    try (Connection conn = getConnection(); Table table = conn.getTable(tableName)) {
-      try (RegionLocator l = conn.getRegionLocator(tableName)) {
-        assertEquals(splitRows.length + 1, l.getAllRegionLocations().size());
-      }
-      // Insert one row each region
-      int insertNum = splitRows.length;
-      for (byte[] splitRow : splitRows) {
-        Put put = new Put(splitRow);
-        put.addColumn(FAMILY, QUALIFIER, VALUE);
-        table.put(put);
-      }
-
-      // scan forward
-      try (ResultScanner scanner = table.getScanner(new Scan())) {
-        int count = 0;
-        for (Result r : scanner) {
-          assertFalse(r.isEmpty());
-          count++;
-        }
-        assertEquals(insertNum, count);
-      }
-
-      // scan backward
-      Scan scan = new Scan();
-      scan.setReversed(true);
-      try (ResultScanner scanner = table.getScanner(scan)) {
-        int count = 0;
-        byte[] lastRow = null;
-        for (Result r : scanner) {
-          assertFalse(r.isEmpty());
-          count++;
-          byte[] thisRow = r.getRow();
-          if (lastRow != null) {
-            assertTrue(Bytes.compareTo(thisRow, lastRow) < 0, "Error scan order, last row= "
-              + Bytes.toString(lastRow) + ",this row=" + Bytes.toString(thisRow));
-          }
-          lastRow = thisRow;
-        }
-        assertEquals(insertNum, count);
-      }
-    }
-  }
-
-  /**
-   * Tests reversed scan under multi regions
-   */
-  @TestTemplate
-  public void testSmallReversedScanUnderMultiRegions() throws Exception {
-    // Test Initialization.
-    byte[][] splitRows = new byte[][] { Bytes.toBytes("000"), Bytes.toBytes("002"),
-      Bytes.toBytes("004"), Bytes.toBytes("006"), Bytes.toBytes("008"), Bytes.toBytes("010") };
-    TEST_UTIL.createTable(tableName, FAMILY, splitRows);
-    TEST_UTIL.waitUntilAllRegionsAssigned(tableName);
-    try (Connection conn = getConnection(); Table table = conn.getTable(tableName)) {
-      try (RegionLocator l = conn.getRegionLocator(tableName)) {
-        assertEquals(splitRows.length + 1, l.getAllRegionLocations().size());
-      }
-      for (byte[] splitRow : splitRows) {
-        Put put = new Put(splitRow);
-        put.addColumn(FAMILY, QUALIFIER, VALUE);
-        table.put(put);
-
-        byte[] nextRow = Bytes.copy(splitRow);
-        nextRow[nextRow.length - 1]++;
-
-        put = new Put(nextRow);
-        put.addColumn(FAMILY, QUALIFIER, VALUE);
-        table.put(put);
-      }
-
-      // scan forward
-      try (ResultScanner scanner = table.getScanner(new Scan())) {
-        int count = 0;
-        for (Result r : scanner) {
-          assertTrue(!r.isEmpty());
-          count++;
-        }
-        assertEquals(12, count);
-      }
-
-      reverseScanTest(table, ReadType.STREAM);
-      reverseScanTest(table, ReadType.PREAD);
-      reverseScanTest(table, ReadType.DEFAULT);
-    }
-  }
-
-  private void reverseScanTest(Table table, ReadType readType) throws IOException {
-    // scan backward
-    Scan scan = new Scan();
-    scan.setReversed(true);
-    try (ResultScanner scanner = table.getScanner(scan)) {
-      int count = 0;
-      byte[] lastRow = null;
-      for (Result r : scanner) {
-        assertTrue(!r.isEmpty());
-        count++;
-        byte[] thisRow = r.getRow();
-        if (lastRow != null) {
-          assertTrue(Bytes.compareTo(thisRow, lastRow) < 0, "Error scan order, last row= "
-            + Bytes.toString(lastRow) + ",this row=" + Bytes.toString(thisRow));
-        }
-        lastRow = thisRow;
-      }
-      assertEquals(12, count);
-    }
-
-    scan = new Scan();
-    scan.setReadType(readType);
-    scan.setReversed(true);
-    scan.withStartRow(Bytes.toBytes("002"));
-    try (ResultScanner scanner = table.getScanner(scan)) {
-      int count = 0;
-      byte[] lastRow = null;
-      for (Result r : scanner) {
-        assertTrue(!r.isEmpty());
-        count++;
-        byte[] thisRow = r.getRow();
-        if (lastRow != null) {
-          assertTrue(Bytes.compareTo(thisRow, lastRow) < 0, "Error scan order, last row= "
-            + Bytes.toString(lastRow) + ",this row=" + Bytes.toString(thisRow));
-        }
-        lastRow = thisRow;
-      }
-      assertEquals(3, count); // 000 001 002
-    }
-
-    scan = new Scan();
-    scan.setReadType(readType);
-    scan.setReversed(true);
-    scan.withStartRow(Bytes.toBytes("002"));
-    scan.withStopRow(Bytes.toBytes("000"));
-    try (ResultScanner scanner = table.getScanner(scan)) {
-      int count = 0;
-      byte[] lastRow = null;
-      for (Result r : scanner) {
-        assertFalse(r.isEmpty());
-        count++;
-        byte[] thisRow = r.getRow();
-        if (lastRow != null) {
-          assertTrue(Bytes.compareTo(thisRow, lastRow) < 0, "Error scan order, last row= "
-            + Bytes.toString(lastRow) + ",this row=" + Bytes.toString(thisRow));
-        }
-        lastRow = thisRow;
-      }
-      assertEquals(2, count); // 001 002
-    }
-
-    scan = new Scan();
-    scan.setReadType(readType);
-    scan.setReversed(true);
-    scan.withStartRow(Bytes.toBytes("001"));
-    try (ResultScanner scanner = table.getScanner(scan)) {
-      int count = 0;
-      byte[] lastRow = null;
-      for (Result r : scanner) {
-        assertFalse(r.isEmpty());
-        count++;
-        byte[] thisRow = r.getRow();
-        if (lastRow != null) {
-          assertTrue(Bytes.compareTo(thisRow, lastRow) < 0, "Error scan order, last row= "
-            + Bytes.toString(lastRow) + ",this row=" + Bytes.toString(thisRow));
-        }
-        lastRow = thisRow;
-      }
-      assertEquals(2, count); // 000 001
-    }
-
-    scan = new Scan();
-    scan.setReadType(readType);
-    scan.setReversed(true);
-    scan.withStartRow(Bytes.toBytes("000"));
-    try (ResultScanner scanner = table.getScanner(scan)) {
-      int count = 0;
-      byte[] lastRow = null;
-      for (Result r : scanner) {
-        assertFalse(r.isEmpty());
-        count++;
-        byte[] thisRow = r.getRow();
-        if (lastRow != null) {
-          assertTrue(Bytes.compareTo(thisRow, lastRow) < 0, "Error scan order, last row= "
-            + Bytes.toString(lastRow) + ",this row=" + Bytes.toString(thisRow));
-        }
-        lastRow = thisRow;
-      }
-      assertEquals(1, count); // 000
-    }
-
-    scan = new Scan();
-    scan.setReadType(readType);
-    scan.setReversed(true);
-    scan.withStartRow(Bytes.toBytes("006"));
-    scan.withStopRow(Bytes.toBytes("002"));
-    try (ResultScanner scanner = table.getScanner(scan)) {
-      int count = 0;
-      byte[] lastRow = null;
-      for (Result r : scanner) {
-        assertFalse(r.isEmpty());
-        count++;
-        byte[] thisRow = r.getRow();
-        if (lastRow != null) {
-          assertTrue(Bytes.compareTo(thisRow, lastRow) < 0, "Error scan order, last row= "
-            + Bytes.toString(lastRow) + ",this row=" + Bytes.toString(thisRow));
-        }
-        lastRow = thisRow;
-      }
-      assertEquals(4, count); // 003 004 005 006
     }
   }
